backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
distro==1.9.0
fastapi==0.135.1
garminconnect==0.2.38
garth==0.5.21
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
oauthlib==3.3.1
openai==2.26.0
playwright==1.58.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
